feat: add upload button in privatekey also
Closes #1098
