+++ apps/extension/src/ui/components/import-components/ImportTabs.tsx
@@ -73,6 +73,7 @@ const ImportTabs = ({
   const [addressFound, setAddressFound] = useState(true);
   const [newKey, setKeyNew] = useState(true);
   const [isLogin, setIsLogin] = useState(false);
+  const [keystoreJson, setKeystoreJson] = useState<string>('');
   const usewallet = useWallet();
   useEffect(() => {
     const checkIsBooted = async () => {
@@ -234,6 +235,7 @@ const ImportTabs = ({
           onImport={handleImport}
           setPk={setPk}
           isSignLoading={isSignLoading}
+          initialJson={keystoreJson}
         />
       </TabPanel>
       <TabPanel value={selectedTab} index={2}>
@@ -254,6 +256,8 @@ const ImportTabs = ({
           onImport={handleImport}
           setPk={setPk}
           isSignLoading={isSignLoading}
+          onSwitchToKeystoreTab={() => setSelectedTab(1)}
+          onSetKeystoreJson={(json) => setKeystoreJson(json)}
         />
       </TabPanel>
       <TabPanel value={selectedTab} index={4}>
